fix: Generation JPA 전환
