+++ src/main/kotlin/co/yappuworld/operation/client/application/AdminUserOperationService.kt
@@ -6,14 +6,14 @@ import co.yappuworld.operation.client.dto.request.AdminGenerationPageRequest
 import co.yappuworld.operation.client.dto.request.AdminGenerationRegisterRequest
 import co.yappuworld.operation.client.dto.response.AdminGenerationActiveUpdateResponse
 import co.yappuworld.operation.client.dto.response.AdminGenerationResponse
-import co.yappuworld.operation.infrastructure.GenerationRepository
+import co.yappuworld.operation.infrastructure.GenerationJpaRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 class AdminUserOperationService(
     private val generationActiveStateManager: GenerationActiveStateManager,
-    private val generationRepository: GenerationRepository
+    private val generationRepository: GenerationJpaRepository
 ) {
 
     @Transactional(readOnly = true)
