--- src/main/kotlin/co/yappuworld/operation/infrastructure/GenerationRepository.kt
@@ -1,18 +1,0 @@
-package co.yappuworld.operation.infrastructure
-
-import co.yappuworld.operation.domain.Generation
-import org.springframework.data.repository.CrudRepository
-import org.springframework.data.repository.PagingAndSortingRepository
-
-interface GenerationRepository :
-    CrudRepository<Generation, Int>,
-    PagingAndSortingRepository<Generation, Int> {
-
-    fun existsGenerationByIsActiveIsTrue(): Boolean
-
-    fun getGenerationOrNullByIsActiveIsTrue(): Generation?
-
-    fun findAllByIsActiveIsTrue(): List<Generation>
-
-    fun findAllByValueIn(values: List<Int>): List<Generation>
-}
