+++ src/test/kotlin/co/yappuworld/schedule/application/ScheduleServiceSessionProgressPhaseTest.kt
@@ -1,6 +1,6 @@
 package co.yappuworld.schedule.application
 
-import co.yappuworld.operation.infrastructure.GenerationRepository
+import co.yappuworld.operation.infrastructure.GenerationJpaRepository
 import co.yappuworld.schedule.client.application.ScheduleService
 import co.yappuworld.schedule.domain.SessionEntity
 import co.yappuworld.schedule.domain.SessionProgressPhase.DONE
@@ -19,7 +19,7 @@ import kotlin.test.assertEquals
 
 class ScheduleServiceSessionProgressPhaseTest {
 
-    private val generationRepository = mockk<GenerationRepository>()
+    private val generationRepository = mockk<GenerationJpaRepository>()
     private val scheduleRepository = mockk<ScheduleJpaRepository>()
     private val scheduleService = ScheduleService(generationRepository, scheduleRepository)
