+++ src/test/kotlin/co/yappuworld/support/fixture/operation/OperationFixture.kt
@@ -1,6 +1,6 @@
 package co.yappuworld.support.fixture.operation
 
-import co.yappuworld.operation.domain.Generation
+import co.yappuworld.operation.domain.GenerationEntity
 import java.time.LocalDate
 
 object OperationFixture {
@@ -10,8 +10,8 @@ object OperationFixture {
         startDate: LocalDate? = LocalDate.of(2025, 11, 13),
         endDate: LocalDate? = LocalDate.of(2025, 3, 8),
         isActive: Boolean = false
-    ): Generation {
-        val generation = Generation(
+    ): GenerationEntity {
+        val generation = GenerationEntity(
             value = value,
             startDate = startDate,
             endDate = endDate
